Don't do the filtering ourselves, thats why kdelibs has widgets for that.
Also port to use QTreeWidget instead of QListWidget, the latter has a bug which
breaks scrolling when some items are hidden (bugreport to TT is on its way).

So far I didn't notice any speed issues with the amount of debug area's kde
currently has (about 600), but QTreeWidget is in general slower with a lot of
items (quite noticeable with 10000) so as soon as QListWidget is fixed
QTreeWidget should be replaced again by it.

I'll add the issue number from TT as soon as I receive it as a comment to this
code

svn path=/trunk/KDE/kdebase/runtime/; revision=815978