KStandardDirs: special case for xdgdata-apps
For resourceDirs("xdgdata-apps"), we do not use
installPath("xdgdata-apps") as the fallback (which is probably a
subdirectory of CMAKE_INSTALL_PREFIX/share/applications/), but instead
use CMAKE_INSTALL_PREFIX/share/applications/ directly.

In a sensible installation, this should make no difference, because if
XDG_DATA_DIRS is set to include CMAKE_INSTALL_PREFIX/share, and that
directory contains an applications directory, the result of
installPath("xdgdata-apps") is never used.  However, this will (a) act
as a bit of a safety blanket for people who set XDG_DATA_DIRS wrong and
(b) allow the tests to pass on Jenkins, which installs kde4support to
its own directory (which then does not have a share/applications
subdirectory).

REVIEW: 114697