Link to KDEWin using PUBLIC on the kdelibs4support target.
It used to be using a global include_directories and link_libraries.
Now we include and link only on the kdelibs4support target and mark both
include dirs and library as PUBLIC. Stuff like mode_t is part of the
public interface of kdelibs4support, so apps using kdelibs4support need
to get kdewin headers.