Encode the URIs which end up in DTD files
The URI need to be encoded, because some valid characters for
filenames are not valid according RFC 2396.
Easy way to trigger the issue: when the path contains spaces,
as it happens on MacOSX builds.

The function is duplicated from the matching functions in kdoctools
because:
- this module will go away in the not far future;
- the function is not generic enough to be used outside (yet)

CHANGELOG: Encode invalid chars in paths used in XML DTDs, avoiding errors
REVIEW: 120649