Call newInstance from the child on first invocation
For the first invocation, call newInstance ourselves instead of relying on the parent process to do it.

With the new threading patches in qt dbus the current code races as the parent may end up calling newInstance before the child has set up the object that handles it.

konsole starts fine with this + Qt 5.6 dbus threading patches

REVIEW: 124856