Fix crash when WICD is not available
When WICD is not available, message is a QDBusMessage::ErrorMessage which has
exactly one argument (the error message) so none of the existing checks would
catch it. The code then tried to get a QDBusArgument from the message arguments
but gets an invalid one because the argument in fact is a QString. Invalid
QDBusArgument does not allow read nor write, and invokes qFatal() in
QDBusArgumentPrivate::checkRead() when we try to deserialize it (when Qt is built
in debug mode).