Provide a fallback MAXPATHLEN if not defined
The ideal fix here would be call realpath() without a fixed-sized
buffer, making use of the POSIX.1-2008 behaviour (where NULL for the
second parameter makes the return value as newly allocated buffer).

OTOH, since this is supposed to be compatibility code and the lack of
MAXPATHLEN does not affect most of the Unix platforms, provide a
fallback definition as less invasive fix.