Fix bug in kfiledialog.cpp that causes crashing when native widgets are used.
This patch makes kfiledialog functions that did not already do so test the value of d->native before referencing d->w. This is important because if d->native is non-null then d->w will be null and referencing it cause an immediate crash. This was the case on Windows builds of okular when the file chooser was opened.

BUGS: 364086
Differential Revision: https://phabricator.kde.org/D2075